==> scripts/one_offs/7.1-compare_UKB_EBB_snp_overlap.ipynb <==
goal: calculate the snp overlap between the ukb main array and the estonian biobank main array


import os
os.getcwd()
os.chdir("/mnt/sdg/robert/deepwidegp/ARD-DeepWideGP/")

os.chdir('/mnt/sdg/robert/deepwidegp/ARD-DeepWideGP')
import pandas as pd

#load the GSA 3.0 (estonian base array)
estonian = pd.read_csv("data/1-inputs/SNP_array_overlap/GSA-24v3-0_A1_estonian_grch37.csv")



#up here you need to navigate to the folder wit the ukb bim files or it wont' work. 
os.chdir('/mnt/sdg/robert/deepwidegp/ARD-DeepWideGP/data/1-inputs/SNP_array_overlap/ukb_bim')
import glob
import pandas as pd

bim_cols = ["chrom", "snp", "cm", "pos", "a1", "a2"]

bim_files = glob.glob("*.bim")  # adjust path if needed

dfs = []
for f in bim_files:
    df = pd.read_csv(
        f,
        sep=r"\s+",
        header=None,
        names=bim_cols,
        dtype={
            "chrom": str,
            "snp": str,
            "cm": float,
            "pos": int,
            "a1": str,
            "a2": str,
        },
    )
    dfs.append(df)

ukb_bim = pd.concat(dfs, ignore_index=True)
print(len(ukb_bim), "variants in combined BIM")


#save the ukb bim df to csv
#os.chdir("/mnt/sdg/robert/deepwidegp/ARD-DeepWideGP/")
#ukb_bim.to_csv("data/1-inputs/SNP_array_overlap/UKB_SNP_array.csv", index=False)

ukb_bim_mapping = ukb_bim.copy()
ukb_bim_mapping['chrom:pos'] = ukb_bim_mapping['chrom'] + ':' + ukb_bim_mapping['pos'].astype(str)

estonian_mapping = estonian.copy()
estonian_mapping['chrom:pos'] = estonian_mapping['Chr'].astype(str) + ':' + estonian_mapping['MapInfo'].round(0).astype('Int64').astype(str)


ukb_bim_mapping.head()

estonian_mapping.head()

#find the overlap
overlap = pd.merge(ukb_bim_mapping, estonian_mapping, on='chrom:pos', how='inner')
print(len(overlap), "overlapping variants between UKB and Estonian array")
print(estonian_mapping['chrom:pos'].nunique(), "unqie variants in Estonian array")
print(ukb_bim_mapping['chrom:pos'].nunique(), "unique variants in UKB BIM files")
